Workaround to make Fay works when using stack nix integration
When using stack with nix enable, ghc and ghc-pkg might not be
under .stack... something but in the nix store (/nix...).
So we need to use the stack wrapper to be in the correct nix
environment.

However, this pause a problem, as we are alread somehow in
a nix environment (and stack forbid it using the STACK_IN_NIX_SHELL
environment variable). We disable it temporatily and it seem to work.